demo_project/requirements.txt
tox
pytest
pytest-django
pytest-cov
coverage
